feat(#39): 메인페이지 캡슐 조회 API 구성

+++ src/main/kotlin/com/yapp/lettie/api/timecapsule/service/reader/TimeCapsuleLikeReader.kt
@@ -29,7 +29,7 @@ class TimeCapsuleLikeReader(
     }
 
     @Transactional(readOnly = true)
-    fun getLikeCount(capsuleId: Long): Long {
+    fun getLikeCount(capsuleId: Long): Int {
         return likeRepository.countByTimeCapsuleIdAndIsLikedTrue(capsuleId)
     }
 }
