+++ src/main/kotlin/com/yapp/lettie/domain/timecapsule/repository/TimeCapsuleLikeRepository.kt
@@ -10,7 +10,7 @@ interface TimeCapsuleLikeRepository : JpaRepository<TimeCapsuleLike, Long> {
         capsule: Long,
     ): TimeCapsuleLike?
 
-    fun countByTimeCapsuleIdAndIsLikedTrue(capsuleId: Long): Long
+    fun countByTimeCapsuleIdAndIsLikedTrue(capsuleId: Long): Int
 
     fun findByUserAndIsLikedTrue(user: User): List<TimeCapsuleLike>
 }
